requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
pytz==2023.3
google-generativeai==0.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
aiofiles==23.2.1
python-dateutil==2.8.2